pytorch_algorithm thread affinity
this patch integrates OpenMP thread affinity settings using TECA's MPI
aware CPU core partitioner. OpenMP's affinity controls are via
environment variable. With torch these take affect upon import. The code
was refactored to defer torch import to after the affinity settings were
made.  If any relevant OMP_<X> variables are detected the affinity
settings are skipped.
