+++ test/python/test_deeplab_ar_detect.py
@@ -37,7 +37,6 @@ ar_detect = teca_deeplab_ar_detect.New()
 ar_detect.set_input_connection(cf_reader.get_output_port())
 ar_detect.set_verbose(vrb)
 ar_detect.set_ivt_variable(water_vapor_var)
-ar_detect.set_threads_per_core(1)
 ar_detect.set_thread_pool_size(n_threads)
 ar_detect.load_model(deeplab_model)
